refactor(DC-704): group startup guards into Startup/Guards namespace
